requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
jinja2==3.1.4
sqlalchemy==2.0.46
aiosqlite==0.22.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
